requirements.txt
fastapi
uvicorn[standard]


langchain
langchain-core
langchain-community
langchain-huggingface
langchain-pinecone
langgraph
langsmith
langchainhub
langchain-text-splitters


openai
google-generativeai
langchain-google-genai
langchain-groq


sentence-transformers
transformers
tiktoken


pinecone


pypdf


python-dotenv
requests
aiofiles
pathlib
wikipedia


dvc


streamlit
flask

ipykernel

-e .
